+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes for Multi Translator
 
+## 2.8.0 - 2024-11-15
+
+### Added
+
+- setting: save as draft, to always create a new draft
+
+### Changed
+
+- moved settings to database
+
 ## 2.7.3 - 2024-11-14
 
 ### Fixed
@@ -101,6 +111,16 @@
 
 ### Updated
 - Craft 5 support
+
+## 1.13.0 - 2024-11-15
+
+### Added
+
+- setting: save as draft, to always create a new draft
+
+### Changed
+
+- moved settings to database
 
 ## 1.12.2 - 2024-10-30